refactor: 이동 EntityFetcher
EntityFetcher를 common.util 패키지로 이동
