+++ src/main/java/inxj/newsfeed/like/service/CommentLikeService.java
@@ -1,6 +1,7 @@
 package inxj.newsfeed.like.service;
 
 import inxj.newsfeed.comment.entity.Comment;
+import inxj.newsfeed.common.util.EntityFetcher;
 import inxj.newsfeed.exception.CustomException;
 import inxj.newsfeed.exception.ErrorCode;
 import inxj.newsfeed.like.entity.CommentLike;
